+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### Current HEAD:
+* Custom orbital reordering for MOLCAS interface
+
+#### Version 1.8.2 LTS update (2016-10-26):
+* Fix PyCheMPS2 build
+
 #### Version 1.8.1 LTS update (2016-10-25):
 * MPS checkpoint for MOLCAS interface
 * Fiedler order checkpoint for MOLCAS interface